Fix bug where files were rewritten w/o newlines
When a file was being rewritten when deleting a block from the queue,
the blocks in the text file were not being separated by newlines, which
will cause bugs when reading back in those blocks, so now they are.

Also fixed a bug where one of Seng's send messages were inside a for
loop, rather than outside.