Add script for making data pack
Takes in a bunch of file paths which will have to be manually added, and
then creates a 'data_pack' directory, which has all these files in, in
the same path directories, but relative to data_pack, rather than
relative to seng. This should then be hopefully easily-enough loaded
into a cloned copy of seng.