Feat/data scripts (#4)
* Add script for making data pack

Takes in a bunch of file paths which will have to be manually added, and
then creates a 'data_pack' directory, which has all these files in, in
the same path directories, but relative to data_pack, rather than
relative to seng. This should then be hopefully easily-enough loaded
into a cloned copy of seng.

* Update the help message

* Correct mistake in the help string

I had not included a word or two.

* Remove unnecessary nesting

I had an if/else statement that wasn't needed, more so the 'else' part.
Removed it to make things less nested than needed.

* Update .gitignore to include the data_pack

data_packs should not be pushed to remote.

* Write the load_data_pack script

Given a data pack, files should now be able to be loaded in, and the
user can choose what to overwrite etc.. i.e. 'This file already exists.
Would you like to replace it with the (new) one from the data pack?'

* Remove unnecessary asterisks from .gitignore

Just leaving a trailing forward slash on the end of directory names will
already exclude the directory and its contents I believe. The asterisk
is thus redundant.