Feat/better welcome message (#20)
* Delete redundant function

* Send welcome block list as a single string

Stops Seng from sending lots of little messages, which will eventually
cause delays due to limit throttling.

* Add a 'send block' method

Reducing repeated code. Method takes in the data for the block to send,
and also the channel to send it in.

* Fix bug where files were rewritten w/o newlines

When a file was being rewritten when deleting a block from the queue,
the blocks in the text file were not being separated by newlines, which
will cause bugs when reading back in those blocks, so now they are.

Also fixed a bug where one of Seng's send messages were inside a for
loop, rather than outside.

* Shorten some function/command names

I basically just abbreviated 'welcome block' to 'wb'. Should make the
interface a bit less type-heavy.

* Include every file in server_specific

.gitignore was not properly ignoring all files in server_specific; files
that were within the welcome_block directory were not being properly
ignored. Fixed this by just removing the wildcard it seems.

* Add doctstrings to welcome message commands

Hopefully this will make things a bit more readable when reading through
this code again.

* Add processes constant

Seng should keep track of which users are using which of Seng's
processes, as some processes may only be 'user-safe' with a single user.
We track this dictionary of 'task: user_id' in a bot constant called
'proceses'.

* Add welcome_block directory and files

These files are necessary and should be created

* Add initial processes code

Method which prints out all processes and their associated members, and
then added in some testing lines for one of the welcome block functions.

* Add asyncless validation methods

To be used in 'check' commands for wait_for, as the check commands are
not async.

* Add process validation and check validation

Only applied to a single Welcome Message command so far.

* Add validation to make_wb command

Hopefully that's all the validation required. Nextt step will be adding
the ability to \'cancel\'.

* Add the option to cancel the process

* Put the cancel command into a function

This will hopefully improve readability, as there's less repeated code.
Overall, the code for this welcome command is pretty messy. Perhaps the
Welcome class could also inherit from some kind of supporting class
which contains process information.

* Add processes and cancel to preview_wb

* Fix a couple typos and add more description

* Add error check to m_preview_wb

If a user enters a block which doesn't exist, Seng will be cool with it

* Add process and cancel checks to relevant method

A process is a command that has the user interact with Seng further. So
all the Welcome Message relevant commands should now have these process
checks implemented. Furthermore, users can now cancel any process
they're in, whenever they like.. otherwise they'd have to complete
processes that they don't want to.

* Conform to PEP8

* Rearranged some code for better readability

Some methods should be next to each other, and check function should be
placed not annoying between two strings.

* Error check to m_add_wb_to_queue

Now we check if the block exists before adding it to the queue

* Add confirmation check to publishing wm

Users should not accidentally use this command, otherwise the welcome
message will be published publicly, unintentionally.

* Fix bug where user not removed from processes

Forgot to add it in

* Conform to PEP8

* Add a remove welcome block function

Welcome blocks can now be deleted from storage

* Add help information for welcome block commands

* Fix bug where 'self' needed to be removed

* Fix bug where cancel method was not being awaited