Fix bug with seeing draft welcome message
Bug was that if someone renames a block, then the queue might not be
able to find that file. We could have any renames also apply to the
queue, but if a moderator renames a file to something else, and then
renames a new block back to the original file, then it may be
counter-intuitive.