Feat/welcome message v2 (#46)
* Improve comments

* Add a couple more comments

* Send actual message if no welcome blocks to list

If there are no welcome blocks in storage, then Seng does not send any
message. In fact, I think an error occurs due to Seng trying to send an
empty string. So now, if the string is empty, Seng  will instead send a
'No blocks etc.' message.

* Add command to clear welcome block queue

Also added info to the welcome blocks help section, and fixed a typo
there that I noticed.

* Add insert welcome block command

* Add welcome block edit command

* Add duplicate welcome block command

* Add rename welcome block command

* Conform to pycodestyle

* Split help string into two

It was too long for Discord to send

* Fix bug where ctx was not passed properly

Instead, I was passing ctx.author.id

* Change the output method of m_view_wb_queue

Instead of sending one block at a time, all blocks are now sent at once.
If there are no blocks, a 'No blocks' string is sent instead.

* Add ctx to get_block_name method calls

It was missing but required!

* Amend typo in get_block_name

Extra space was needed

* Fix bug where author_id was being used

Needed to be the replaced ctx.author.id

* Fix typo spelling mistake

* Fix bug where self was missing on method call

* Fix bug where 'await' was missing

* Fix bug where variable name was wrong

* Finish fixing that bug from last commit

Turned out there were more mistakes where I had not changed all the
variables.

* Fix bug where 'self' was missing from method calls

* Fix bug where I was accidentally using .content

* Fix bug where block not being added if queue empty

It's because if 'blocks' is empty, then there's nothing to iterate over,
so the counter conditional is never ran.

* Fix bug where .json was not appended to src/dst

Must be appended

* Fix bugs where rename_wb was not working

* Fix bug with seeing draft welcome message

Bug was that if someone renames a block, then the queue might not be
able to find that file. We could have any renames also apply to the
queue, but if a moderator renames a file to something else, and then
renames a new block back to the original file, then it may be
counter-intuitive.

* Fix bugs for edit command

* Fix a couple of final welcome message bugs

* Conform to pycodestyle